refactor(diagrams): unify ValidationError/Warning/Result into one shared module
Final piece of the 2026-05-21 pre-release should-fix batch. The cross-module
divergence concern from the orchestrator's review:

  > ValidationError / ValidationResult are re-declared in every validate.ts,
  > error-code vocabularies differ (activities uses ACT-*, others use generic
  > codes), and only process-blueprint guards null elements. A shared
  > validation-types.ts would prevent the next instance of the blocker-1
  > class.

This PR addresses the FIRST half: one canonical type triple. The error-code
vocabulary divergence and per-element null guards are notation-by-notation
concerns and stay as today — touching them broadly would balloon this PR.

## Change

New file `packages/diagrams/src/validation-types.ts` exports the canonical
`ValidationError` / `ValidationWarning` / `ValidationResult` triple. Every
notation's validate.ts (and the two notations that route types through
their own types.ts — goals, activities) now imports the shape from there
instead of declaring its own copy.

Public-surface compatibility:

- Notations that exposed PREFIXED names (`FGCA…`, `Activity…`,
  `ProcessBlueprint…`) keep those names as type aliases of the shared
  shape. Consumers that imported the prefixed names continue to compile
  unchanged.
- Goals re-exports the unprefixed names through its own index.ts (existing
  public surface) — that re-export now resolves to the shared declaration.
- The package root index.ts also exports the shared module directly, so
  consumers can import `ValidationError` from `@transitrix/diagrams` without
  going through a specific notation.

## Width of the change

| File | Before | After |
|---|---|---|
| `validation-types.ts` (new) | — | canonical types |
| `goals/types.ts` | local interfaces | re-export from shared |
| `activities/types.ts` | local prefixed interfaces | type aliases of shared |
| `fgca/validate.ts` | local prefixed interfaces | type aliases of shared |
| `capability-map/validate.ts` | local interfaces | re-export from shared |
| `process-map/validate.ts` | local interfaces | re-export from shared |
| `applications/validate.ts` | local interfaces | re-export from shared |
| `products/validate.ts` | local interfaces | re-export from shared |
| `scenarios/validate.ts` | local interfaces | re-export from shared |
| `process-blueprint/validate.ts` | local interfaces | re-export from shared |
| `index.ts` | per-notation re-exports | + direct shared re-export |

Six of the previous declarations had a slightly narrower `{code, message}`
shape; the shared one carries the optional `path?: string` that goals,
fgca, activities, and process-blueprint already used. Widening from
narrower to wider is structurally compatible — any existing object literal
with just `{code, message}` is still assignable.

## Tests

No new tests — this is a type-only refactor, no runtime behaviour changes.
The existing suite confirms compatibility:

- diagrams: 307 / 307 passing.
- core: 247 / 247 passing.
- `tsc --noEmit` clean across root, `packages/diagrams`, and `extension`.

The third file (`extension`) is the load-bearing check here: the Studio
preview that imports types from `@transitrix/diagrams/process-blueprint`
(and the inlined-copy previews that mirror the package types) compile
without changes.
