+++ packages/diagrams/src/scenarios/validate.ts
@@ -1,12 +1,7 @@
 import type { ScenarioStatus, FactorRelevance } from './types.js';
+import type { ValidationError, ValidationWarning, ValidationResult } from '../validation-types.js';
 
-export interface ValidationError { code: string; message: string; }
-export interface ValidationWarning { code: string; message: string; }
-export interface ValidationResult {
-  valid: boolean;
-  errors: ValidationError[];
-  warnings: ValidationWarning[];
-}
+export type { ValidationError, ValidationWarning, ValidationResult };
 
 const VALID_STATUSES = new Set<ScenarioStatus>(['Draft', 'Active', 'Archived']);
 const VALID_RELEVANCE = new Set<FactorRelevance>(['High', 'Medium', 'Low']);
